update libsodium
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@38182 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/libsodium/src/libsodium/crypto_hash/sha512/cp/hash_sha512_cp.c
@@ -172,10 +172,10 @@ static const uint8_t PAD[128] = {
 static void
 SHA512_Pad(crypto_hash_sha512_state *state, uint64_t tmp64[80 + 8])
 {
-    uint64_t r;
-    uint64_t i;
+    unsigned int r;
+    unsigned int i;
 
-    r = (state->count[1] >> 3) & 0x7f;
+    r = (unsigned int) ((state->count[1] >> 3) & 0x7f);
     if (r < 112) {
         for (i = 0; i < 112 - r; i++) {
             state->buf[r + i] = PAD[i];
@@ -222,9 +222,11 @@ crypto_hash_sha512_update(crypto_hash_sha512_state *state,
 
     bitlen[1] = ((uint64_t) inlen) << 3;
     bitlen[0] = ((uint64_t) inlen) >> 61;
+    /* LCOV_EXCL_START */
     if ((state->count[1] += bitlen[1]) < bitlen[1]) {
         state->count[0]++;
     }
+    /* LCOV_EXCL_STOP */
     state->count[0] += bitlen[0];
     if (inlen < 128 - r) {
         for (i = 0; i < inlen; i++) {
